Refactor Pin Indicator widgets
