+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* Initial release
+* Initial release: Pinpad and PinIndicator (24 animations)
