+++ lib/pin_ui.dart
@@ -5,7 +5,6 @@ export 'src/indicator/exceptions/controller_not_initialized_exception.dart';
 export 'src/indicator/models/implementations.dart'
     hide PinAnimationImplementation;
 export 'src/indicator/widgets/pin_indicator.dart';
-export 'src/indicator/widgets/pin_indicator_builder.dart';
 export 'src/indicator/widgets/pin_indicator_item.dart';
 export 'src/pinpad/models/extra_key.dart';
 export 'src/pinpad/pinpad.dart';
